Apply some frontend restructuring
