+++ frontend/lib/src/components/elements/ArrowVegaLiteChart/arrowUtils.test.ts
